Fix: Move unit tests into dedicated namespace
