+++ Makefile
@@ -9,4 +9,4 @@ cs: composer
 	vendor/bin/php-cs-fixer fix --config=.php_cs --verbose --diff
 
 test: composer
-	vendor/bin/phpunit --configuration phpunit.xml --columns max
+	vendor/bin/phpunit --configuration=test/Unit/phpunit.xml
