+++ test/Unit/phpunit.xml
@@ -1,11 +1,12 @@
 <phpunit
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.8/phpunit.xsd"
-    bootstrap="vendor/autoload.php"
+    bootstrap="../../vendor/autoload.php"
     beStrictAboutChangesToGlobalState="true"
     beStrictAboutOutputDuringTests="true"
     beStrictAboutTestsThatDoNotTestAnything="true"
     colors="true"
+    columns="max"
     convertErrorsToExceptions="true"
     convertNoticesToExceptions="true"
     convertWarningsToExceptions="true"
@@ -17,12 +18,12 @@
 >
     <testsuites>
         <testsuite name="Unit Tests">
-            <directory>test</directory>
+            <directory>.</directory>
         </testsuite>
     </testsuites>
     <filter>
         <whitelist addUncoveredFilesFromWhitelist="true" processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">src</directory>
+            <directory suffix=".php">../../src</directory>
         </whitelist>
     </filter>
 </phpunit>
